Add interfield comparison rule only if supported
Previously a rule was added which attempted to check for a root user accessing
files owned by other users in /home/ by using interfield comparison between the
audit userid and the file owner userid.

Support for interfield comparison (-C option) was only added to auditctl in
version 2.2, so distributions such as Ubuntu 12.04 do not support it.

On these systems, auditctl failed to load the rule and processing stopped, so
some rules would not be loaded, including custom rules defined by the user.

We now only install this rule on:

- Ubuntu 14.04 or more recent versions
- RedHat / CentOS running 6.3 or more recent versions