Use beaker-rspec to test example code from README
In order to test that the module works on both current Ubuntu LTS releases, we
beaker with the vagrant driver to stand up a 12.04 and a 14.04 VM, run through
the example code from the README and ensure that the audit rule has been
correctly installed and appears in the output of auditctl -l.

Since this can't be run from travis, we're relying on comitters / mergers to do
this themselves locally.