Remove rspec-system test
There was a basic rspec-system test, but it doesn't appear to have worked for
some time, since it doesn't install the concat module onto the system under
test.

There is a slightly more useful beaker test for this module now and
rspec-system's github page indicates that it is obsolete and beaker should be
used instead.